requirements.txt
python-docx>=1.1.0
lxml>=4.9.0
streamlit>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
